refactor(tui): move agent history activity helpers
Extract compact agent launch rendering, activity-group rendering, and agent id parsing from history.rs into history/agent_activity.rs while preserving live and transcript behavior.

Refs #3308.
